pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-linux-admin"
version = "3.0.0"
authors = [
    { name="Your Name", email="you@example.com" },
]
description = "A modern, post-quantum encrypted Telegram bot for Linux server administration with 2026 standards."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Topic :: System :: Systems Administration",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "python-telegram-bot[ext]>=21.0",
    "asyncssh>=2.14",
    "tenacity>=8.2",
    "async-timeout>=4.0",
    "requests>=2.31",
    "cryptography>=42.0",
    "psutil>=5.9; extra == 'monitoring'",
]
dependencies-optional = [
    "psutil>=5.9",
]

[project.optional-dependencies]
monitoring = ["psutil>=5.9"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.4",
    "black>=24.0",
]

[project.scripts]
tla-bot = "src.main:main"
tla-bot-update = "src.updater:main"
